scenarios/batch/code/src/requirements.txt
setuptools
wheel
psutil
azure.batch
azure.identity
azure.common
azure.keyvault
applicationinsights
pandas
numpy
scipy
redis